--- docs/docs/codes/keymap-upgrader.mdx
@@ -1,21 +1,0 @@
----
-title: Keymap Upgrader
-sidebar_label: Keymap Upgrader
-hide_title: true
-hide_table_of_contents: true
----
-
-# Keymap Upgrader
-
-Many codes have been renamed to be more consistent with each other.
-Paste the contents of a `.keymap` file below to upgrade all deprecated codes to their replacements.
-
-Hover your mouse over the upgraded keymap and click the `Copy` button to copy it to your clipboard.
-
-You will likely need to realign columns in the upgraded keymap. The upgrader also does not handle
-codes inside a `#define`, so you will need to update those manually using
-[this list of deprecated codes and replacements](https://github.com/zmkfirmware/zmk/blob/main/docs/src/data/keymap-upgrade.js).
-
-import KeymapUpgrader from "@site/src/components/KeymapUpgrader/index";
-
-<KeymapUpgrader />
