+++ docs/docusaurus.config.js
@@ -56,6 +56,11 @@ module.exports = {
           position: "left",
         },
         {
+          to: "keymap-upgrader",
+          label: "Keymap Upgrader",
+          position: "left",
+        },
+        {
           href: "https://github.com/zmkfirmware/zmk",
           label: "GitHub",
           position: "right",
