+++ docs/src/pages/keymap-upgrader.mdx
@@ -1,0 +1,20 @@
+---
+title: Keymap Upgrader
+sidebar_label: Keymap Upgrader
+hide_title: true
+hide_table_of_contents: true
+---
+
+# Keymap Upgrader
+
+Some behaviors, key codes, and other features have been renamed to be more consistent with each other. This tool will upgrade most deprecated features to their replacements.
+
+Paste the contents of a `.keymap` file below. Then, hover your mouse over the upgraded keymap and click the `Copy` button in the upper-right corner to copy it to your clipboard.
+
+You will likely need to realign columns in the upgraded keymap. The upgrader also does not handle
+codes inside a `#define`, so you will need to update those manually using
+[this list of deprecated codes and replacements](https://github.com/zmkfirmware/zmk/blob/main/docs/src/keymap-upgrade/keycodes.ts).
+
+import KeymapUpgrader from "@site/src/components/KeymapUpgrader/index";
+
+<KeymapUpgrader />
